requirements.txt
pip~=23.3.2
cryptography~=41.0.7
redis~=5.0.1
amqp~=5.2.0
vine~=5.1.0
Django~=5.0.1
kombu~=5.3.4
billiard~=4.2.0
celery~=5.3.6
click~=8.1.7
python-dateutil~=2.8.2
asgiref~=3.7.2
sqlparse~=0.4.4
colorama~=0.4.6
requests~=2.31.0
wcwidth~=0.2.12
cffi~=1.16.0
six~=1.16.0
djangorestframework~=3.14.0
setuptools~=65.5.0
django-ckeditor~=6.7.0